requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
line-bot-sdk==3.14.2
google-generativeai==0.8.3
redis==5.2.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
